requirements.txt
python-telegram-bot
cryptography
python-nmap
quart
pyngrok
requests
schedule
aiohttp
python-dotenv

